+++ gwcelery/tests/test_sentry.py
@@ -6,8 +6,7 @@ from urllib.parse import urlparse, urlunparse
 
 import pytest
 
-from .. import __version__
-from .. import sentry
+from .. import __version__, app, sentry
 
 
 @pytest.fixture
@@ -38,4 +37,4 @@ def test_sentry_configure(mock_celery_integration, mock_sdk_init,
     mock_celery_integration.assert_called_once_with()
     mock_sdk_init.assert_called_once_with(
         dsn, integrations=[mock_celery_integration.return_value],
-        release=__version__)
+        environment=app.conf['sentry_environment'], release=__version__)